Update to new engine version + flake

+++ src/util.rs
@@ -7,7 +7,7 @@ pub trait SoundExt {
 impl SoundExt for geng::Sound {
     fn play_random_pitch(&self) {
         let mut effect = self.effect();
-        const OFFSET: f64 = 0.2;
+        const OFFSET: f32 = 0.2;
         effect.set_speed(thread_rng().gen_range(1.0 - OFFSET..1.0 + OFFSET));
         effect.play();
     }
